+++ source/TR-181/middle_layer_src/cosa_ip_dml.c
@@ -2457,7 +2457,7 @@ IPv4Address_GetParamUlongValue
 #endif /* !_COSA_BCM_MIPS_ */
         {
             /* TBC -- this should be from a SBAPI call */
-            *puLong = CosaUtilGetIfAddr(pIPInterface->Cfg.LinkName);
+            *puLong = (ULONG) CosaUtilGetIfAddr(pIPInterface->Cfg.LinkName);
         }
          
          return TRUE;
